Put s3_bucket_and_s3 in conftest.py
